+++ database/migrations/0001_01_01_000000_create_merchant_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->rememberToken();
-            $table->integer('amount');
+            $table->float('amount', 2);
             $table->timestamps();
         });
